Move some functions to new block.c file
