+++ src/input.h
@@ -1,7 +1,7 @@
-#include "structs.h"
-
 #ifndef _input_h_
 #define _input_h_
+
+#include "structs.h"
 
 void handle_mouse_input(Model *model);
 
@@ -19,24 +19,7 @@ int hit_test(
 
 int hit_test_face(Player *player, int *x, int *y, int *z, int *face, Model *model);
 
-void unset_sign(int x, int y, int z, Model *model);
-
-void set_light(int p, int q, int x, int y, int z, int w);
-
-void _set_block(int p, int q, int x, int y, int z, int w, int dirty, Model *model);
-
-void set_block(int x, int y, int z, int w, Model *model);
-
-void record_block(int x, int y, int z, int w, Model *model);
-
-int get_block(int x, int y, int z, Model *model);
-
 void on_left_click(Model *model);
-
-int player_intersects_block(
-    int height,
-    float x, float y, float z,
-    int hx, int hy, int hz);
 
 void on_right_click(Model *model);
 
@@ -60,22 +43,6 @@ void _set_sign(
 void add_message(const char *text, Model *model);
 
 void login(Model *model);
-
-void builder_block(int x, int y, int z, int w, Model *model);
-
-void copy(Model *model);
-
-void paste(Model *model);
-
-void array(Block *b1, Block *b2, int xc, int yc, int zc, Model *model);
-
-void cube(Block *b1, Block *b2, int fill, Model *model);
-
-void sphere(Block *center, int radius, int fill, int fx, int fy, int fz, Model *model);
-
-void cylinder(Block *b1, Block *b2, int radius, int fill, Model *model);
-
-void tree(Block *block, Model *model);
 
 void parse_command(const char *buffer, int forward, Model *model);
